make it work right when light-locker isn't even installed

+++ hooks/customise.sh
@@ -104,7 +104,9 @@ chroot ${rootdir} apt -y remove linux-image-4.9.0-3-amd64 || true
 chroot ${rootdir} chmod 755 /
 
 chroot ${rootdir} mkdir -p /etc/skel/.config/autostart
+if [ -e ${rootdir}/etc/xdg/autostart/light-locker.desktop ]; then
 chroot ${rootdir} desktop-file-install --dir=/etc/skel/.config/autostart --set-key=Hidden --set-value=true /etc/xdg/autostart/light-locker.desktop
+fi
 
 echo "blacklist bochs-drm" > $rootdir/etc/modprobe.d/qemu-blacklist.conf
